+++ src/ImageOptimizer/WrapperOptimizer.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace ImageOptimizer;
+
+interface WrapperOptimizer extends Optimizer {
+    public function unwrap(): Optimizer;
+}